requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]==3.11
scikit-learn
dill
pyaml
mlflow
##-e .